# Faster Parquet Bloom Writer