# Which issue does this PR close?

<!--
We generally require a GitHub issue to be filed for all bug fixes and enhancements and this helps us generate change logs for our releases. You can link an issue to this PR using the GitHub syntax. For example `Closes #123` indicates that this PR will close issue #123.
-->

Part of https://github.com/apache/arrow-rs/issues/3320.

# Rationale for this change
 
<!--
Why are you proposing this change? If this is already explained clearly in the issue then this section is not needed.
Explaining clearly why changes are proposed helps reviewers understand your changes and offer better suggestions for fixes.
-->


```
$ RUSTFLAGS="-C target-cpu=native" cargo bench --bench arrow_writer -- --baseline master "bloom filter"

write_batch primitive/4096 values primitive with bloom filter
                        time:   [16.561 ms 16.848 ms 17.135 ms]
                        thrpt:  [10.302 MiB/s 10.477 MiB/s 10.659 MiB/s]
                 change:
                        time:   [-11.831% -10.178% -8.6583%] (p = 0.00 < 0.05)
                        thrpt:  [+9.4790% +11.331% +13.419%]
                        Performance has improved.
write_batch primitive/4096 values primitive non-null with bloom filter
                        time:   [5.4278 ms 5.4894 ms 5.5546 ms]
                        thrpt:  [31.164 MiB/s 31.534 MiB/s 31.892 MiB/s]
                 change:
                        time:   [-70.261% -69.885% -69.508%] (p = 0.00 < 0.05)
                        thrpt:  [+227.96% +232.06% +236.26%]
                        Performance has improved.
Found 1 outliers among 100 measurements (1.00%)
  1 (1.00%) high mild
Benchmarking write_batch primitive/4096 values string with bloom filter: Warming up for 3.0000 s
Warning: Unable to complete 100 samples in 5.0s. You may wish to increase target time to 7.6s, enable flat sampling, or reduce sample count to 50.
write_batch primitive/4096 values string with bloom filter
                        time:   [1.4824 ms 1.4851 ms 1.4879 ms]
                        thrpt:  [53.521 MiB/s 53.623 MiB/s 53.722 MiB/s]
                 change:
                        time:   [-21.422% -21.052% -20.634%] (p = 0.00 < 0.05)
                        thrpt:  [+25.998% +26.666% +27.263%]
                        Performance has improved.
Found 3 outliers among 100 measurements (3.00%)
  2 (2.00%) high mild
  1 (1.00%) high severe
write_batch primitive/4096 values string dictionary with bloom filter
                        time:   [675.68 µs 677.72 µs 679.75 µs]
                        thrpt:  [70.823 MiB/s 71.035 MiB/s 71.249 MiB/s]
                 change:
                        time:   [-21.076% -20.658% -20.301%] (p = 0.00 < 0.05)
                        thrpt:  [+25.472% +26.037% +26.704%]
                        Performance has improved.
Found 18 outliers among 100 measurements (18.00%)
  16 (16.00%) low mild
  2 (2.00%) high severe
Benchmarking write_batch primitive/4096 values string non-null with bloom filter: Warming up for 3.0000 s
Warning: Unable to complete 100 samples in 5.0s. You may wish to increase target time to 7.7s, enable flat sampling, or reduce sample count to 50.
write_batch primitive/4096 values string non-null with bloom filter
                        time:   [1.5209 ms 1.5253 ms 1.5297 ms]
                        thrpt:  [51.422 MiB/s 51.569 MiB/s 51.719 MiB/s]
                 change:
                        time:   [-29.848% -29.673% -29.495%] (p = 0.00 < 0.05)
                        thrpt:  [+41.835% +42.192% +42.548%]
                        Performance has improved.
```


# What changes are included in this PR?

<!--
There is no need to duplicate the description in the issue here but it is sometimes worth providing a summary of the individual changes in this PR.
-->

# Are there any user-facing changes?


<!--
If there are user-facing changes then we may require documentation to be updated before approving the PR.
-->

<!---
If there are any breaking changes to public APIs, please add the `breaking change` label.
-->